Expanding the DTO to handle blank JSON fields
Specifically, Go handles *incoming* fields fine (-ish, and we may still
need to handle incoming blanks, but that'll be more work...), but
outgoing JSON exhibits the unfornate behavior that all fields must have
a value... This commit adds the idea of a "presence" map to DTO objects,
and only serialzies out fields which are present. This does mean that
the natural approach to creating DTOs (i.e. Type{Field: "value"}) won't
work - the Field in the example won't be serialized. This behavior is
about as suprising as Go's normal JSON handling